requirements.txt
beautifulsoup4==4.11.1
Django==4.0.5
gunicorn==20.1.0
psycopg2==2.9.3
dj-database-url
django-heroku
djangorestframework
djangorestframework-simplejwt
django-cors-headers
boto3
